Update import paths and wire up the initial package.json
This commit fixes all imports paths that were broken after
merging the monorepo packages togther. It then wires up a new and
fully functional single package.json for the entire project, which
replaces the package.json's that were in the monorepo packages.
All package.json scripts included are functional, which means
tests are passing and bundles are being built.

Note: `npm run deploy` has intentionally been left out for now
since our deployment process is going to change quite
significantly.
