+++ .vscode/settings.json
@@ -4,16 +4,5 @@
   "editor.rulers": [80],
   "files.trimTrailingWhitespace": true,
   "files.insertFinalNewline": true,
-  "files.exclude": {
-    "**/.git": true,
-    "**/.DS_Store": true,
-    "**/.rpt2_cache": true,
-    "node_modules": true,
-    "test-lib": true,
-    "lib": true,
-    "dist": true,
-    "coverage": true,
-    "npm": true
-  },
   "typescript.tsdk": "../node_modules/typescript/lib"
 }
