+++ config/jest.config.js
@@ -1,21 +1,19 @@
 module.exports = {
+  rootDir: '..',
   transform: {
     '.(ts|tsx)': 'ts-jest',
   },
-
   globals: {
     'ts-jest': {
       diagnostics: false,
     },
   },
-
   moduleFileExtensions: ['ts', 'tsx', 'js', 'json'],
   testURL: 'http://localhost',
-
   testMatch: ['<rootDir>/src/**/__tests__/**/*.ts'],
   testPathIgnorePatterns: [
     '/node_modules/',
-    '/lib/',
-    '<rootDir>/lib/',
+    '/lib/'
   ],
+  setupFiles: ['<rootDir>/src/config/jest/setup.ts'],
 };
