--- config/tsconfig.base.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "noImplicitAny": true,
-    "noUnusedParameters": false,
-    "noUnusedLocals": true,
-    "skipLibCheck": true,
-    "moduleResolution": "node",
-    "importHelpers": true,
-    "removeComments": true,
-    "sourceMap": true,
-    "declaration": true,
-    "declarationMap": true,
-    "target": "es5",
-    "module": "commonjs",
-    "esModuleInterop": true
-  }
-}
