--- jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  rootDir: '.',
-  projects: ['<rootDir>/packages/*'],
-};
