+++ package.json
@@ -1,57 +1,63 @@
 {
-  "name": "apollo-client-monorepo",
-  "private": true,
+  "name": "@apollo/client",
+  "version": "3.0.0-alpha.0",
+  "description": "A fully-featured caching GraphQL client.",
+  "author": "opensource@apollographql.com",
   "license": "MIT",
+  "main": "./lib/bundle.cjs.js",
+  "module": "./lib/bundle.esm.js",
+  "sideEffects": [
+    "./lib/fixPolyfills.js"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/apollographql/apollo-client.git"
+  },
+  "bugs": {
+    "url": "https://github.com/apollographql/apollo-client/issues"
+  },
+  "homepage": "https://www.apollographql.com",
   "scripts": {
-    "postinstall": "lerna exec -- npm install --package-lock=false && lerna run prepare",
-    "diff": "check-if-folder-contents-changed-in-git-commit-range",
-    "build": "lerna run build",
-    "test": "jest --verbose",
-    "test-ci": "npm run coverage -- --ci --maxWorkers=2 --reporters=default --reporters=jest-junit",
-    "benchmark": "cd packages/apollo-client && npm run benchmark",
-    "lint": "lerna run lint",
-    "filesize": "lerna run build && bundlesize",
-    "coverage": "jest --verbose --coverage",
-    "coverage:upload": "codecov",
-    "danger": "danger run --verbose --dangerfile=./config/dangerfile.ts",
-    "deploy": "lerna publish -m \"chore: Publish\" && cd packages/apollo-client && npm run deploy",
-    "clean": "find . -name \"node_modules\" -exec rm -rf '{}' + && find . -name \"lib\" -exec rm -rf '{}' +"
+    "prebuild": "npm run clean",
+    "build": "npx tsc",
+    "postbuild": "npm run bundle",
+    "watch": "npx tsc-watch --onSuccess \"npm run postbuild\"",
+    "clean": "npx rimraf -r lib coverage",
+    "postinstall": "npm run build",
+    "test": "jest --config ./config/jest.config.js",
+    "test:ci": "npm run coverage -- --ci --maxWorkers=2 --reporters=default --reporters=jest-junit",
+    "test:watch": "jest --config ./config/jest.config.js --watch",
+    "bundle": "npx rollup -c ./config/rollup.config.js",
+    "coverage": "npx jest --config ./config/jest.config.js --verbose --coverage",
+    "coverage:upload": "npx codecov",
+    "bundlesize": "npm run build && bundlesize"
   },
   "bundlesize": [
     {
-      "name": "apollo-cache",
-      "path": "./packages/apollo-cache/lib/bundle.cjs.min.js",
-      "maxSize": "825 B"
-    },
-    {
-      "name": "apollo-cache-inmemory",
-      "path": "./packages/apollo-cache-inmemory/lib/bundle.cjs.min.js",
-      "maxSize": "4.3 kB"
-    },
-    {
       "name": "apollo-client",
-      "path": "./packages/apollo-client/lib/bundle.cjs.min.js",
-      "maxSize": "9.9 kB"
-    },
-    {
-      "name": "apollo-utilities",
-      "path": "./packages/apollo-utilities/lib/bundle.cjs.min.js",
-      "maxSize": "4.25 kB"
+      "path": "./lib/bundle.cjs.min.js",
+      "maxSize": "11.8 kB"
     }
   ],
   "dependencies": {
-    "apollo-cache": "file:packages/apollo-cache",
-    "apollo-cache-inmemory": "file:packages/apollo-cache-inmemory",
-    "apollo-client": "file:packages/apollo-client",
-    "apollo-utilities": "file:packages/apollo-utilities"
+    "@types/zen-observable": "^0.8.0",
+    "@wry/equality": "^0.1.9",
+    "apollo-link": "^1.2.13",
+    "fast-json-stable-stringify": "^2.0.0",
+    "optimism": "^0.10.2",
+    "symbol-observable": "^1.2.0",
+    "ts-invariant": "^0.4.4",
+    "tslib": "^1.10.0",
+    "zen-observable": "^0.8.14"
+  },
+  "peerDependencies": {
+    "graphql": "^14.0.0"
   },
   "devDependencies": {
     "@babel/core": "7.5.5",
     "@babel/plugin-transform-modules-commonjs": "7.5.0",
     "@babel/plugin-transform-modules-umd": "7.2.0",
     "@condenast/bundlesize": "0.18.1",
-    "@octokit/rest": "16.28.7",
-    "@types/benchmark": "1.0.31",
     "@types/graphql": "14.2.3",
     "@types/isomorphic-fetch": "0.0.35",
     "@types/jest": "24.0.16",
@@ -59,21 +65,18 @@
     "@types/node": "12.6.9",
     "@types/react": "16.8.24",
     "@types/react-dom": "16.8.5",
-    "benchmark": "2.1.4",
-    "check-if-folder-contents-changed-in-git-commit-range": "1.0.0",
     "codecov": "3.5.0",
-    "danger": "7.1.4",
     "fetch-mock": "7.3.9",
     "graphql": "14.4.2",
     "graphql-tag": "2.10.1",
     "isomorphic-fetch": "2.2.1",
     "jest": "23.6.0",
     "jest-junit": "5.2.0",
-    "lerna": "3.16.4",
     "lodash": "4.17.15",
     "prop-types": "15.7.2",
     "react": "16.8.6",
     "react-dom": "16.8.6",
+    "rimraf": "^3.0.0",
     "rollup": "1.17.0",
     "rollup-plugin-invariant": "0.5.6",
     "rollup-plugin-local-resolve": "1.0.7",
@@ -84,41 +87,7 @@
     "rxjs": "6.5.2",
     "ts-jest": "23.10.5",
     "tsc-watch": "2.2.1",
-    "tslib": "1.10.0",
-    "tslint": "5.18.0",
     "typescript": "3.5.3",
     "uglify-js": "3.6.0"
-  },
-  "renovate": {
-    "extends": [
-      ":pinOnlyDevDependencies"
-    ],
-    "semanticCommits": true,
-    "timezone": "America/Los_Angeles",
-    "schedule": [
-      "after 10pm and before 5am on every weekday"
-    ],
-    "rebaseStalePrs": true,
-    "prCreation": "not-pending",
-    "automerge": "minor",
-    "labels": [
-      "dependencies"
-    ],
-    "assignees": [
-      "@hwillson"
-    ],
-    "reviewers": [
-      "@hwillson"
-    ],
-    "pathRules": [
-      {
-        "paths": [
-          "docs/package.json"
-        ],
-        "extends": [
-          "apollo-docs"
-        ]
-      }
-    ]
   }
 }
