+++ src/__mocks__/mockQueryManager.ts
@@ -3,7 +3,7 @@ import { QueryManager } from '../core/QueryManager';
 import { mockSingleLink, MockedResponse } from './mockLinks';
 
 import { DataStore } from '../data/store';
-import { InMemoryCache } from 'apollo-cache-inmemory';
+import { InMemoryCache } from '../cache/inmemory';
 
 // Helper method for the tests that construct a query manager out of a
 // a list of mocked responses for a mocked network interface.
