+++ src/__tests__/fetchMore.ts
@@ -1,8 +1,8 @@
-import { InMemoryCache } from 'apollo-cache-inmemory';
 import { assign, cloneDeep } from 'lodash';
 import gql from 'graphql-tag';
 
 import { mockSingleLink } from '../__mocks__/mockLinks';
+import { InMemoryCache } from '../cache/inmemory';
 import ApolloClient, { NetworkStatus, ObservableQuery } from '../';
 
 describe('updateQuery on a simple query', () => {
