+++ src/__tests__/local-state/export.ts
@@ -1,9 +1,9 @@
 import gql from 'graphql-tag';
-
-import ApolloClient from '../..';
-import { InMemoryCache } from 'apollo-cache-inmemory';
 import { ApolloLink, Observable } from 'apollo-link';
 import { print } from 'graphql/language/printer';
+
+import ApolloClient from '../..';
+import { InMemoryCache } from '../../cache/inmemory';
 
 describe('@client @export tests', () => {
   it(
