+++ src/__tests__/mutationResults.ts
@@ -1,11 +1,10 @@
 import { ApolloLink, Observable } from 'apollo-link';
 import { cloneDeep } from 'lodash';
 import gql from 'graphql-tag';
-import { InMemoryCache } from 'apollo-cache-inmemory';
 
 import { mockSingleLink } from '../__mocks__/mockLinks';
 import ApolloClient from '..';
-
+import { InMemoryCache } from '../cache/inmemory';
 import { Subscription } from '../util/Observable';
 import { withWarning } from '../util/wrap';
