+++ src/__tests__/subscribeToMore.ts
@@ -1,13 +1,11 @@
 import gql from 'graphql-tag';
 import { ApolloLink, Operation } from 'apollo-link';
-import { InMemoryCache } from 'apollo-cache-inmemory';
-import { stripSymbols } from 'apollo-utilities';
 
 import { DocumentNode, OperationDefinitionNode } from 'graphql';
-
 import { mockSingleLink, mockObservableLink } from '../__mocks__/mockLinks';
-
 import ApolloClient from '../';
+import { InMemoryCache } from '../cache/inmemory';
+import { stripSymbols } from '../utilities';
 
 const isSub = (operation: Operation) =>
   (operation.query as DocumentNode).definitions
