+++ src/data/queries.ts
@@ -1,6 +1,7 @@
 import { DocumentNode, GraphQLError, ExecutionResult } from 'graphql';
-import { isEqual } from 'apollo-utilities';
 import { invariant } from 'ts-invariant';
+
+import { isEqual } from '../utilities';
 import { NetworkStatus } from '../core/networkStatus';
 import { isNonEmptyArray } from '../util/arrays';
