+++ src/utilities/__tests__/fragments.ts
@@ -5,7 +5,12 @@ import { disableFragmentWarnings } from 'graphql-tag';
 // Turn off warnings for repeated fragment names
 disableFragmentWarnings();
 
-import { getFragmentQueryDocument } from '../fragments';
+import {
+  getFragmentQueryDocument,
+  createFragmentMap,
+  FragmentMap
+} from '../fragments';
+import { getFragmentDefinitions } from '../getFromAST';
 
 describe('getFragmentQueryDocument', () => {
   it('will throw an error if there is an operation', () => {
@@ -324,4 +329,27 @@ describe('getFragmentQueryDocument', () => {
       `),
     );
   });
+});
+
+it('should create the fragment map correctly', () => {
+  const fragments = getFragmentDefinitions(gql`
+    fragment authorDetails on Author {
+      firstName
+      lastName
+    }
+
+    fragment moreAuthorDetails on Author {
+      address
+    }
+  `);
+  const fragmentMap = createFragmentMap(fragments);
+  const expectedTable: FragmentMap = {
+    authorDetails: fragments[0],
+    moreAuthorDetails: fragments[1],
+  };
+  expect(fragmentMap).toEqual(expectedTable);
+});
+
+it('should return an empty fragment map if passed undefined argument', () => {
+  expect(createFragmentMap(undefined)).toEqual({});
 });
