+++ src/utilities/__tests__/getFromAST.ts
@@ -7,8 +7,6 @@ import {
   getFragmentDefinitions,
   getQueryDefinition,
   getMutationDefinition,
-  createFragmentMap,
-  FragmentMap,
   getDefaultValues,
   getOperationName,
 } from '../getFromAST';
@@ -188,29 +186,6 @@ describe('AST utility functions', () => {
       .definitions[0] as OperationDefinitionNode;
     const actualResult = getMutationDefinition(mutationWithFragments);
     expect(print(actualResult)).toEqual(print(expectedResult));
-  });
-
-  it('should create the fragment map correctly', () => {
-    const fragments = getFragmentDefinitions(gql`
-      fragment authorDetails on Author {
-        firstName
-        lastName
-      }
-
-      fragment moreAuthorDetails on Author {
-        address
-      }
-    `);
-    const fragmentMap = createFragmentMap(fragments);
-    const expectedTable: FragmentMap = {
-      authorDetails: fragments[0],
-      moreAuthorDetails: fragments[1],
-    };
-    expect(fragmentMap).toEqual(expectedTable);
-  });
-
-  it('should return an empty fragment map if passed undefined argument', () => {
-    expect(createFragmentMap(undefined)).toEqual({});
   });
 
   it('should get the operation name out of a query', () => {
