+++ src/utilities/fragments.ts
@@ -1,4 +1,9 @@
-import { DocumentNode, FragmentDefinitionNode } from 'graphql';
+import {
+  DocumentNode,
+  FragmentDefinitionNode,
+  InlineFragmentNode,
+  SelectionNode
+} from 'graphql';
 import { invariant, InvariantError } from 'ts-invariant';
 
 /**
@@ -89,4 +94,40 @@ export function getFragmentQueryDocument(
   };
 
   return query;
+}
+
+/**
+ * This is an interface that describes a map from fragment names to fragment definitions.
+ */
+export interface FragmentMap {
+  [fragmentName: string]: FragmentDefinitionNode;
+}
+
+// Utility function that takes a list of fragment definitions and makes a hash out of them
+// that maps the name of the fragment to the fragment definition.
+export function createFragmentMap(
+  fragments: FragmentDefinitionNode[] = [],
+): FragmentMap {
+  const symTable: FragmentMap = {};
+  fragments.forEach(fragment => {
+    symTable[fragment.name.value] = fragment;
+  });
+  return symTable;
+}
+
+export function getFragmentFromSelection(
+  selection: SelectionNode,
+  fragmentMap: FragmentMap,
+): InlineFragmentNode | FragmentDefinitionNode | null {
+  switch (selection.kind) {
+    case 'InlineFragment':
+      return selection;
+    case 'FragmentSpread': {
+      const fragment = fragmentMap && fragmentMap[selection.name.value];
+      invariant(fragment, `No fragment named ${selection.name.value}.`);
+      return fragment;
+    }
+    default:
+      return null;
+  }
 }
