+++ src/utilities/getFromAST.ts
@@ -3,8 +3,6 @@ import {
   OperationDefinitionNode,
   FragmentDefinitionNode,
   ValueNode,
-  InlineFragmentNode,
-  SelectionNode,
 } from 'graphql';
 
 import { invariant, InvariantError } from 'ts-invariant';
@@ -167,42 +165,6 @@ export function getMainDefinition(
   throw new InvariantError(
     'Expected a parsed GraphQL query with a query, mutation, subscription, or a fragment.',
   );
-}
-
-/**
- * This is an interface that describes a map from fragment names to fragment definitions.
- */
-export interface FragmentMap {
-  [fragmentName: string]: FragmentDefinitionNode;
-}
-
-// Utility function that takes a list of fragment definitions and makes a hash out of them
-// that maps the name of the fragment to the fragment definition.
-export function createFragmentMap(
-  fragments: FragmentDefinitionNode[] = [],
-): FragmentMap {
-  const symTable: FragmentMap = {};
-  fragments.forEach(fragment => {
-    symTable[fragment.name.value] = fragment;
-  });
-  return symTable;
-}
-
-export function getFragmentFromSelection(
-  selection: SelectionNode,
-  fragmentMap: FragmentMap,
-): InlineFragmentNode | FragmentDefinitionNode | null {
-  switch (selection.kind) {
-    case 'InlineFragment':
-      return selection;
-    case 'FragmentSpread': {
-      const fragment = fragmentMap && fragmentMap[selection.name.value];
-      invariant(fragment, `No fragment named ${selection.name.value}.`);
-      return fragment;
-    }
-    default:
-      return null;
-  }
 }
 
 export function getDefaultValues(
