+++ tsconfig.json
@@ -1,9 +1,24 @@
 {
-  "extends": "./config/tsconfig.base.json",
-  "references": [
-    { "path": "packages/apollo-cache" },
-    { "path": "packages/apollo-cache-inmemory" },
-    { "path": "packages/apollo-client" },
-    { "path": "packages/apollo-utilities" }
+  "compilerOptions": {
+    "noImplicitAny": true,
+    "noUnusedParameters": false,
+    "noUnusedLocals": true,
+    "skipLibCheck": true,
+    "moduleResolution": "node",
+    "importHelpers": true,
+    "removeComments": true,
+    "sourceMap": true,
+    "declaration": true,
+    "declarationMap": true,
+    "target": "es5",
+    "module": "es2015",
+    "esModuleInterop": true,
+    "outDir": "./lib",
+    "lib": ["es2015", "dom"]
+  },
+  "include": ["src/**/*.ts"],
+  "exclude": [
+    "src/**/__tests__/**/*",
+    "src/**/__mocks__/**/*"
   ]
 }
